Refactor: Split and organize JavaScript into utils/article.js
