+++ app/javascript/application.js
@@ -9,58 +9,4 @@ Rails.start();
 import 'trix';
 import '@rails/actiontext';
 
-import $ from 'jquery';
-import axios from 'axios';
-
-const csrfToken = document
-  .querySelector('meta[name=csrf-token]')
-  ?.getAttribute('content');
-axios.defaults.headers.common['X-CSRF-Token'] = csrfToken;
-
-const handleHeartDisplay = (hasLiked) => {
-  if (hasLiked) {
-    $('.active-heart').removeClass('hidden');
-  } else {
-    $('.inactive-heart').removeClass('hidden');
-  }
-};
-
-document.addEventListener('turbo:load', () => {
-  const dataset = $(`#article-show`).data();
-  const articleId = dataset.articleId;
-
-  axios.get(`/articles/${articleId}/like`).then((response) => {
-    const hasLiked = response.data.hasLiked;
-    handleHeartDisplay(hasLiked);
-  });
-
-  $('.inactive-heart').on('click', () => {
-    axios
-      .post(`/articles/${articleId}/like`)
-      .then((response) => {
-        if (response.data.status === 'ok') {
-          $('.active-heart').removeClass('hidden');
-          $('.inactive-heart').addClass('hidden');
-        }
-      })
-      .catch((e) => {
-        window.alert('Error');
-        console.log(e);
-      });
-  });
-
-  $('.active-heart').on('click', () => {
-    axios
-      .delete(`/articles/${articleId}/like`)
-      .then((response) => {
-        if (response.data.status === 'ok') {
-          $('.inactive-heart').removeClass('hidden');
-          $('.active-heart').addClass('hidden');
-        }
-      })
-      .catch((e) => {
-        window.alert('Error');
-        console.log(e);
-      });
-  });
-});
+import './utils/article';
